Partially revert df4edffc6d17c3b223322278054cfa9837195270 VoiceAttack needs decimal values rather than doubles. Revise jump calcs to exclude aux fuel reservoir.

+++ DataDefinitions/JumpDetail.cs
@@ -5,14 +5,14 @@ namespace EddiDataDefinitions
     public class JumpDetail
     {
         [PublicAPI]
-        public double distance { get; private set; }
+        public decimal distance { get; private set; }
 
         [PublicAPI]
         public int jumps { get; private set; }
 
         public JumpDetail() { }
 
-        public JumpDetail(double distance, int jumps)
+        public JumpDetail(decimal distance, int jumps)
         {
             this.distance = distance;
             this.jumps = jumps;
